refactor(web): item (#1051)
* content refactor

* fix member type

* fix: link item modal hooks

* fix: link item request modal hooks

* fix: dropdown render hooks

* revert content table

* revert workspace

* revert user member
