+++ web/src/components/molecules/Content/LinkItemModal/hooks.ts
@@ -1,8 +1,16 @@
-import { useCallback } from "react";
+import { useState, useEffect, ChangeEvent, useCallback, useMemo } from "react";
 
+import { TablePaginationConfig } from "@reearth-cms/components/atoms/ProTable";
 import { useIsItemReferencedLazyQuery } from "@reearth-cms/gql/graphql-client-api";
 
-export default () => {
+export default (
+  linkItemModalTotalCount?: number,
+  linkItemModalPage?: number,
+  linkItemModalPageSize?: number,
+  visible?: boolean,
+) => {
+  const [value, setValue] = useState("");
+
   const [checkIfItemIsReferenced, { data }] = useIsItemReferencedLazyQuery({
     fetchPolicy: "no-cache",
   });
@@ -13,8 +21,32 @@ export default () => {
     [checkIfItemIsReferenced],
   );
 
+  const pagination: TablePaginationConfig = useMemo(
+    () => ({
+      showSizeChanger: true,
+      current: linkItemModalPage,
+      total: linkItemModalTotalCount,
+      pageSize: linkItemModalPageSize,
+    }),
+    [linkItemModalPage, linkItemModalTotalCount, linkItemModalPageSize],
+  );
+
+  const handleInput = useCallback((e: ChangeEvent<HTMLInputElement>) => {
+    setValue(e.target.value);
+  }, []);
+
+  useEffect(() => {
+    if (!visible) {
+      setValue("");
+    }
+  }, [visible]);
+
   return {
     isReferenced: data?.isItemReferenced,
+    value,
+    toolbar,
+    pagination,
+    handleInput,
     handleCheckItemReference,
   };
 };
