+++ web/src/components/molecules/Content/LinkItemModal/index.tsx
@@ -1,15 +1,11 @@
 import styled from "@emotion/styled";
-import { useState, useEffect, ChangeEvent, useCallback } from "react";
+import { useState, useCallback, useMemo } from "react";
 
 import Button from "@reearth-cms/components/atoms/Button";
 import Icon from "@reearth-cms/components/atoms/Icon";
 import Input from "@reearth-cms/components/atoms/Input";
 import Modal from "@reearth-cms/components/atoms/Modal";
-import ProTable, {
-  ProColumns,
-  TablePaginationConfig,
-  ListToolBarProps,
-} from "@reearth-cms/components/atoms/ProTable";
+import ProTable, { ProColumns, ListToolBarProps } from "@reearth-cms/components/atoms/ProTable";
 import { useT } from "@reearth-cms/i18n";
 import { dateTimeFormat } from "@reearth-cms/utils/format";
 
@@ -49,115 +45,113 @@ const LinkItemModal: React.FC<Props> = ({
   const [hoveredAssetId, setHoveredItemId] = useState<string>();
   const t = useT();
   const { confirm } = Modal;
-  const { handleCheckItemReference } = useHooks();
-  const [value, setValue] = useState("");
+  const { value, pagination, handleInput, handleCheckItemReference } = useHooks(
+    linkItemModalTotalCount,
+    linkItemModalPage,
+    linkItemModalPageSize,
+    visible,
+  );
 
-  const pagination: TablePaginationConfig = {
-    showSizeChanger: true,
-    current: linkItemModalPage,
-    total: linkItemModalTotalCount,
-    pageSize: linkItemModalPageSize,
-  };
+  const handleClick = useCallback(
+    async (link: boolean, item: FormItem) => {
+      if (!link) {
+        onChange?.("");
+        onLinkItemModalCancel();
+        return;
+      }
 
-  const columns: ProColumns<FormItem>[] = [
-    {
-      title: "",
-      width: 40,
-      render: (_, item) => {
-        const link =
-          (item.id === linkedItem && hoveredAssetId !== item.id) ||
-          (item.id !== linkedItem && hoveredAssetId === item.id);
-        return (
-          <Button
-            type="link"
-            onMouseEnter={() => setHoveredItemId(item.id)}
-            onMouseLeave={() => setHoveredItemId(undefined)}
-            icon={<Icon icon={link ? "linkSolid" : "unlinkSolid"} size={16} />}
-            onClick={async () => {
-              if (!link) {
-                onChange?.("");
-                onLinkItemModalCancel();
-                return;
-              }
+      const response = await handleCheckItemReference(item.id, correspondingFieldId);
+      const isReferenced = response?.data?.isItemReferenced;
 
-              const response = await handleCheckItemReference(item.id, correspondingFieldId);
-              const isReferenced = response?.data?.isItemReferenced;
+      if (isReferenced) {
+        confirm({
+          title: t("This item has been referenced"),
+          content: t(
+            "Are you going to refer to it? The previous reference will be canceled automatically",
+          ),
+          icon: <Icon icon="exclamationCircle" />,
+          onOk() {
+            onChange?.(item.id);
+            onLinkItemModalCancel();
+          },
+        });
+      } else {
+        onChange?.(item.id);
+        onLinkItemModalCancel();
+      }
+    },
+    [confirm, correspondingFieldId, handleCheckItemReference, onChange, onLinkItemModalCancel, t],
+  );
 
-              if (isReferenced) {
-                confirm({
-                  title: t("This item has been referenced"),
-                  content: t(
-                    "Are you going to refer to it? The previous reference will be canceled automatically",
-                  ),
-                  icon: <Icon icon="exclamationCircle" />,
-                  onOk() {
-                    onChange?.(item.id);
-                    onLinkItemModalCancel();
-                  },
-                });
-              } else {
-                onChange?.(item.id);
-                onLinkItemModalCancel();
-              }
-            }}
-          />
-        );
+  const columns: ProColumns<FormItem>[] = useMemo(
+    () => [
+      {
+        title: "",
+        width: 40,
+        render: (_, item) => {
+          const link =
+            (item.id === linkedItem && hoveredAssetId !== item.id) ||
+            (item.id !== linkedItem && hoveredAssetId === item.id);
+          return (
+            <Button
+              type="link"
+              onMouseEnter={() => setHoveredItemId(item.id)}
+              onMouseLeave={() => setHoveredItemId(undefined)}
+              icon={<Icon icon={link ? "linkSolid" : "unlinkSolid"} size={16} />}
+              onClick={() => handleClick(link, item)}
+            />
+          );
+        },
       },
-    },
-    {
-      title: "ID",
-      dataIndex: "id",
-      key: "id",
-      ellipsis: true,
-    },
-    {
-      title: t("Title"),
-      dataIndex: "title",
-      key: "title",
-      ellipsis: true,
-    },
-    {
-      title: t("Created By"),
-      dataIndex: "createdBy",
-      key: "createdBy",
-      ellipsis: true,
-    },
-    {
-      title: t("Created At"),
-      dataIndex: "createdAt",
-      key: "createdAt",
-      ellipsis: true,
-      render: (_text, record) => dateTimeFormat(record.createdAt),
-    },
-  ];
+      {
+        title: "ID",
+        dataIndex: "id",
+        key: "id",
+        ellipsis: true,
+      },
+      {
+        title: t("Title"),
+        dataIndex: "title",
+        key: "title",
+        ellipsis: true,
+      },
+      {
+        title: t("Created By"),
+        dataIndex: "createdBy",
+        key: "createdBy",
+        ellipsis: true,
+      },
+      {
+        title: t("Created At"),
+        dataIndex: "createdAt",
+        key: "createdAt",
+        ellipsis: true,
+        render: (_text, record) => dateTimeFormat(record.createdAt),
+      },
+    ],
+    [t, linkedItem, hoveredAssetId, handleClick],
+  );
 
-  useEffect(() => {
-    if (!visible) {
-      setValue("");
-    }
-  }, [visible]);
-
-  const handleInput = useCallback((e: ChangeEvent<HTMLInputElement>) => {
-    setValue(e.target.value);
-  }, []);
-
-  const toolbar: ListToolBarProps = {
-    search: (
-      <Input.Search
-        allowClear
-        placeholder={t("input search text")}
-        onSearch={(value: string) => {
-          if (value) {
-            onSearchTerm(value);
-          } else {
-            onSearchTerm();
-          }
-        }}
-        value={value}
-        onChange={handleInput}
-      />
-    ),
-  };
+  const toolbar: ListToolBarProps = useMemo(
+    () => ({
+      search: (
+        <Input.Search
+          allowClear
+          placeholder={t("input search text")}
+          onSearch={(value: string) => {
+            if (value) {
+              onSearchTerm(value);
+            } else {
+              onSearchTerm();
+            }
+          }}
+          value={value}
+          onChange={handleInput}
+        />
+      ),
+    }),
+    [t, value, handleInput, onSearchTerm],
+  );
 
   return (
     <Modal
